- add check for the covers of the currently inserted disc to the cover downloader (is checked with the other games in the list) - fix booting wii discs when starting without a HDD (channel mode) - add parental control option to block priiloader override on return to system menu - add prompt asking for downloading new language files when trying to update and no language files exist yet - reduced amount of checks for inserted disc (now every 100ms) - fixed booting gamecube games (datatype bug for mount method) - fix winnertag for gamecube games - add per game nand emu path setting

+++ source/prompts/gameinfo.cpp
@@ -25,7 +25,7 @@
 #include "utils/ShowError.h"
 #include "BoxCover/BoxCover.hpp"
 
-extern int mountMethod;
+extern u8 mountMethod;
 extern struct discHdr *dvdheader;
 
 /****************************************************************************
